+++ template/Cargo.toml
@@ -5,5 +5,5 @@ edition = "2021"
 publish = false
 
 [dependencies]
-uefi = { version = "0.18.0", features = ["exts"] }
+uefi = { version = "0.18.0", features = ["alloc"] }
 uefi-services = "0.15.0"
